+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### Added
+- Documentation for the speed detector
+
 ## v7.1.0
 
 ### Added